Split all of util/FormField into separate modules that can be tree-shaken away
