+++ src/util/forms/getFormError.js
@@ -1,0 +1,10 @@
+import { getIn } from "formik";
+
+const getFormError = (name, touched, errors) => {
+    const error = getIn(errors, name);
+    const touch = getIn(touched, name);
+
+    return touch && error ? error : null;
+};
+
+export default getFormError;
